migrating props lesson to shared components
